Refactored classes to current architecture
