+++ src/main/java/pl/gov/coi/pomocua/ads/assistance/LegalAssistance.java
@@ -1,29 +1,19 @@
-package pl.gov.coi.pomocua.ads.dictionaries.domain;
+package pl.gov.coi.pomocua.ads.assistance;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import lombok.Data;
-import org.hibernate.validator.constraints.Length;
+import lombok.EqualsAndHashCode;
+import org.springframework.data.repository.PagingAndSortingRepository;
+import pl.gov.coi.pomocua.ads.BaseOffer;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
 
 import java.util.List;
 
 import static javax.persistence.EnumType.STRING;
-import static javax.persistence.GenerationType.IDENTITY;
 
-@Data
+@EqualsAndHashCode(callSuper = true)
 @Entity
-@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.PUBLIC_ONLY)
-public class LegalAssistance {
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @NotBlank
-    @Length(max = 80)
-    public String title;
-
+public class LegalAssistance extends BaseOffer {
     @Enumerated(STRING)
     public Mode mode;
 
@@ -40,13 +30,12 @@ public class LegalAssistance {
     @Enumerated(STRING)
     public List<Language> language;
 
-    @NotBlank
-    @Length(max = 1000)
-    public String description;
-
     enum Language {}
 
     enum Mode {}
 
     enum HelpType {}
+}
+
+interface LegalAssistanceRepository  extends PagingAndSortingRepository<LegalAssistance, Long> {
 }
