--- src/main/java/pl/gov/coi/pomocua/ads/dictionaries/api/LegalAssistanceResponse.java
@@ -1,10 +1,0 @@
-package pl.gov.coi.pomocua.ads.dictionaries.api;
-
-import pl.gov.coi.pomocua.ads.dictionaries.domain.LegalAssistance;
-
-import java.util.List;
-
-public class LegalAssistanceResponse {
-    public LegalAssistanceResponse(List<LegalAssistance> legalAssistanceList) {
-    }
-}
