--- src/main/java/pl/gov/coi/pomocua/ads/dictionaries/domain/LegalAssistanceRepository.java
@@ -1,6 +1,0 @@
-package pl.gov.coi.pomocua.ads.dictionaries.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface LegalAssistanceRepository  extends CrudRepository<LegalAssistance, Long> {
-}
